UI layer specification
When a bunch of text is drawn by other resources, it often gets between the menu
text and it's texture/background, and the whole thing can be a mess.

Upside:  It's all on the same layer
Downside: Anything drawn after it, that does not decleare a layer, ends up
messing with it anyway.