Implemented a "hide function" feature
Basically, it's a function that is automagically called every time you hide a
menu using menu.hide()
I found myself implementing convoluted ways to reset some values when a menu is
closed/hidden, and suspect others might be doing the same, so why not make it
easier for everyone.

It gets called with a reference to the menu as it's *only* argument.